refactor(questura): modularize media set migration
